power-saving: add ddr pll toggle access in menu
Major developments:
- Added an option in the ddr menu to turn the ddr pll
on or off.
- Modified options to place, clear and verify patterns in
whole blocks of memory. These functions essentially sweep
through the ddr memory and does whatever it needs to do.
This way we can have a larger scope of the ddr and how it
is affected.

Minor devlopments:
- Some formatting and naming changes also.
- Added LIM-Debug build to git ignore
- The references to the ddr menu refers to the self
refresh in particular. This should be more generic,
thus it was changed to reflect a more generic menu.

Jira-id: ESSBM-254
Signed-off-by: Charles Haden <Charles.Haden@microchip.com>