power-saving: fix state machine
This is an update to the state machine section of the MPFS
HAL example project. The issue before was that if the state
machine was run by the user, it would be impossible to
return back to the WAIT_START state. There was also an
inability to return back to default power states by using
this mode. These updates fix these issues, allowing the
state machine to function correctly and also giving the
option to go back to full power or low power modes.

Jira-id: ESSBM-230
Signed-off-by: Charles Haden <Charles.Haden@microchip.com>