rm fpga_design_config output files
These are output files that are generated when a
build is carried out on a project. They're not
necessary to have within the projects themselves.
These output files can be found in the builds
folder and they mostly serve to highlight the
build config to the user.

They are generated files and are not necessary
within the base project. Thus, this commit removes
these output files and adds a .gitignore change
to make sure these output files don't end up in
any commits in the future.

Jira-id: ESSBM-396
Signed-off-by: Garo Janir <garo.janir@microchip.com>