README: update per disco kit + modify sections
Major developments:
- The README for the example projects has been modified to
provide information about the discovery kit. The projects
that are available on the discovery kit have been noted in
the Supported Devices section. Parts of the document that
talk solely about the icicle kit have been modified to use
a generic "polarfire device" when necessary.
- The table of contents has been updated to correspond with
"Markdown All in One" VSCode extension, which allows the
table of contents to be automatically updated when a user
changes a markdown header.
- More images and graphs, and tables have been added to give
the user a better visual understanding of what is being
explained in the document.
- The overall structure of the document has been changed
to a more streamlined and organised approach to the document
from the user's perspective.
- Any redundant or superfluous information was removed and
more fleshed-out sections. Paragraphs were moved to more
appropriate sections of the document.

Jira-id: ESSBM-402
Signed-off-by: Charles Haden <Charles.Haden@microchip.com>