tree-wide: fix project build names per proposal
1. Fix formatting
Each one the .cproject files in the examples repository
was correctly formatted. Those projects that have been
untouched for a while still have the old formatting which
is a mix of spacing and tabbing.

2. Fix build names
The build names have been changed to remove any ambiguity
on where the program is being stored and executed, and
whether it is a debug or release build. There is a standard
build name format that is being used from now on

3. Affected projects
sys-serv, timer, usb, watchdog, mmuart, qspi,
spi, rtc, ethernet-mac, i2c, can, gpio, mmc, core*
mpfs-hal, user-crypto

NOTE: dma benchmarking projects weren't changed as they
will be changed later

NOTE: This commit does not have any functional affects on
the example project files. It is merely a formatting change
that automatically happens when you click apply in the
project properties settings in SoftConsole.

Jira-id: ESSBM-406
Signed-off-by: Charles Haden <Charles.Haden@microchip.com>